// src/client/zod.gen.ts
export const zPatientManagementBasicInfoRequestWritable = z.object({
    utm: z.enum([
        'app',
        'utility'
    ]),
    appointment_calendar_interval: z.uuid(),
    appointment_booking: z.uuid(),
    national_number: z.string().min(1).max(10),
    name: z.string().min(1).max(255),
    gender: z.enum([
        'Male',
        'Female'
    ]),
    birth_date: z.iso.date(),
    reason: z.optional(z.enum([
        'visit',
        'lab'
    ])),
    phone_number: z.optional(z.union([
        z.string().max(13),
        z.null()
    ])),
    fheight: z.optional(z.union([
        z.number().gte(30).lte(250),
        z.null()
    ])),
    mheight: z.optional(z.union([
        z.number().gte(30).lte(250),
        z.null()
    ])),
    insurance: z.optional(z.nullable(z.enum([
        'T_E',
        'N_M',
        'KH_D',
        'T',
        'O',
        ''
    ])))
});